+++ lib/src/dart/map_extension.dart
@@ -14,9 +14,7 @@ extension MapExtension<T> on Map<T, T> {
   ///print(map.has("id", 1)); // true
   ///print(map.has("id", 2)); // false
   ///```
-  bool has(String key, dynamic value) {
-    return containsKey(key) && this[key] == value;
-  }
+  bool has(String key, dynamic value) => containsKey(key) && this[key] == value;
 
   ///If this map does not contains the given [key]/[value] pair.
   ///
@@ -69,7 +67,7 @@ extension MapExtension<T> on Map<T, T> {
   ///```dart
   ///map.diffKeys(map2)
   ///```
-  Map<T, T> diffKeys(Map<T, T> map) {
+  Map<T, T> diffKeys<K, V>(Map<K, V> map) {
     removeWhere((T key, T value) => map.containsKey(key));
     return this;
   }