+++ lib/src/flutter/extensions/text_color_extension.dart
@@ -1,108 +1,254 @@
 import 'package:flutter/material.dart';
 import 'package:quickly/quickly.dart';
 
-/// Custom color palette
+/// Custom color extension for the Text color. This extension provides getter methods for various commonly used colors in the application.
 extension TextColorExtension on Text {
+  /// Returns Text with Primary color
   Text get kcPrimary => color(FxColor.primary);
+
+  /// Returns Text with Secondary color
   Text get kcSecondary => color(FxColor.secondary);
+
+  /// Returns Text with Success color
   Text get kcSuccess => color(FxColor.success);
+
+  /// Returns Text with Info color
   Text get kcInfo => color(FxColor.info);
+
+  /// Returns Text with Warning color
   Text get kcWarning => color(FxColor.warning);
+
+  /// Returns Text with Danger color
   Text get kcDanger => color(FxColor.danger);
+
+  /// Returns Text with Light color
   Text get kcLight => color(FxColor.light);
+
+  /// Returns Text with Dark color
   Text get kcDark => color(FxColor.dark);
+
+  /// Returns Text with Black color
   Text get black => color(FxColor.black);
+
+  /// Returns Text with White color
   Text get white => color(FxColor.white);
 
-  Text get kcText => color(FxColor.kcText);
-  Text get kcLabel => color(FxColor.kcLabel);
-  Text get kcDisabled => color(FxColor.kcDisabled);
-  Text get kcTextField => color(FxColor.kcTextField);
+  /// Returns Text with Warning Light color
+  Text get kcWarningLight => color(FxColor.kcWarningLight);
 
-  // Status Colors
-  Text get kcWarningStatus => color(FxColor.kcWarningStatus);
-  Text get kcDangerStatus => color(FxColor.kcDangerStatus);
-  Text get kcSuccessStatus => color(FxColor.kcSuccessStatus);
+  /// Returns Text with Danger Light color
+  Text get kcDangerLight => color(FxColor.kcDangerLight);
 
-  // Light Status Colors
-  Text get kcWarningLightStatus => color(FxColor.kcWarningLightStatus);
-  Text get kcDangerLightStatus => color(FxColor.kcDangerLightStatus);
-  Text get kcSuccessLightStatus => color(FxColor.kcSuccessLightStatus);
+  /// Returns Text with Success Light color
+  Text get kcSuccessLight => color(FxColor.kcSuccessLight);
 
+  /// Returns Text with gray50 color
   Text get gray50 => color(FxColor.gray50);
+
+  /// Returns Text with gray100 color
   Text get gray100 => color(FxColor.gray100);
+
+  /// Returns Text with gray200 color
   Text get gray200 => color(FxColor.gray200);
+
+  /// Returns Text with gray300 color
   Text get gray300 => color(FxColor.gray300);
+
+  /// Returns Text with gray400 color
   Text get gray400 => color(FxColor.gray400);
+
+  /// Returns Text with gray500 color
   Text get gray500 => color(FxColor.gray500);
+
+  /// Returns Text with gray600 color
   Text get gray600 => color(FxColor.gray600);
+
+  /// Returns Text with gray700 color
   Text get gray700 => color(FxColor.gray700);
+
+  /// Returns Text with gray800 color
   Text get gray800 => color(FxColor.gray800);
+
+  /// Returns Text with gray900 color
   Text get gray900 => color(FxColor.gray900);
 
+  /// Returns Text with neutral50 color
   Text get neutral50 => color(FxColor.neutral50);
+
+  /// Returns Text with neutral100 color
   Text get neutral100 => color(FxColor.neutral100);
+
+  /// Returns Text with neutral200 color
   Text get neutral200 => color(FxColor.neutral200);
+
+  /// Returns Text with neutral300 color
   Text get neutral300 => color(FxColor.neutral300);
+
+  /// Returns Text with neutral400 color
   Text get neutral400 => color(FxColor.neutral400);
+
+  /// Returns Text with neutral500 color
   Text get neutral500 => color(FxColor.neutral500);
+
+  /// Returns Text with neutral600 color
   Text get neutral600 => color(FxColor.neutral600);
+
+  /// Returns Text with neutral700 color
   Text get neutral700 => color(FxColor.neutral700);
+
+  /// Returns Text with neutral800 color
   Text get neutral800 => color(FxColor.neutral800);
+
+  /// Returns Text with neutral900 color
   Text get neutral900 => color(FxColor.neutral900);
 
+  /// Returns Text with amber50 color
   Text get amber50 => color(FxColor.amber50);
+
+  /// Returns Text with amber100 color
   Text get amber100 => color(FxColor.amber100);
+
+  /// Returns Text with amber200 color
   Text get amber200 => color(FxColor.amber200);
+
+  /// Returns Text with amber300 color
   Text get amber300 => color(FxColor.amber300);
+
+  /// Returns Text with amber400 color
   Text get amber400 => color(FxColor.amber400);
+
+  /// Returns Text with amber500 color
   Text get amber500 => color(FxColor.amber500);
+
+  /// Returns Text with amber600 color
   Text get amber600 => color(FxColor.amber600);
+
+  /// Returns Text with amber700 color
   Text get amber700 => color(FxColor.amber700);
+
+  /// Returns Text with amber800 color
   Text get amber800 => color(FxColor.amber800);
+
+  /// Returns Text with amber900 color
   Text get amber900 => color(FxColor.amber900);
 
+  /// Returns Text with red50 color
   Text get red50 => color(FxColor.red50);
+
+  /// Returns Text with red100 color
   Text get red100 => color(FxColor.red100);
+
+  /// Returns Text with red200 color
   Text get red200 => color(FxColor.red200);
+
+  /// Returns Text with red300 color
   Text get red300 => color(FxColor.red300);
+
+  /// Returns Text with red400 color
   Text get red400 => color(FxColor.red400);
+
+  /// Returns Text with red500 color
   Text get red500 => color(FxColor.red500);
+
+  /// Returns Text with red600 color
   Text get red600 => color(FxColor.red600);
+
+  /// Returns Text with red700 color
   Text get red700 => color(FxColor.red700);
+
+  /// Returns Text with red800 color
   Text get red800 => color(FxColor.red800);
+
+  /// Returns Text with red900 color
   Text get red900 => color(FxColor.red900);
 
+  /// Returns Text with green50 color
   Text get green50 => color(FxColor.green50);
+
+  /// Returns Text with green100 color
   Text get green100 => color(FxColor.green100);
+
+  /// Returns Text with green200 color
   Text get green200 => color(FxColor.green200);
+
+  /// Returns Text with green300 color
   Text get green300 => color(FxColor.green300);
+
+  /// Returns Text with green400 color
   Text get green400 => color(FxColor.green400);
+
+  /// Returns Text with green500 color
   Text get green500 => color(FxColor.green500);
+
+  /// Returns Text with green600 color
   Text get green600 => color(FxColor.green600);
+
+  /// Returns Text with green700 color
   Text get green700 => color(FxColor.green700);
+
+  /// Returns Text with green800 color
   Text get green800 => color(FxColor.green800);
+
+  /// Returns Text with green900 color
   Text get green900 => color(FxColor.green900);
 
+  /// Returns Text with blue50 color
   Text get blue50 => color(FxColor.blue50);
+
+  /// Returns Text with blue100 color
   Text get blue100 => color(FxColor.blue100);
+
+  /// Returns Text with blue200 color
   Text get blue200 => color(FxColor.blue200);
+
+  /// Returns Text with blue300 color
   Text get blue300 => color(FxColor.blue300);
+
+  /// Returns Text with blue400 color
   Text get blue400 => color(FxColor.blue400);
+
+  /// Returns Text with blue500 color
   Text get blue500 => color(FxColor.blue500);
+
+  /// Returns Text with blue600 color
   Text get blue600 => color(FxColor.blue600);
+
+  /// Returns Text with blue700 color
   Text get blue700 => color(FxColor.blue700);
+
+  /// Returns Text with blue800 color
   Text get blue800 => color(FxColor.blue800);
+
+  /// Returns Text with blue900 color
   Text get blue900 => color(FxColor.blue900);
 
+  /// Returns Text with violet50 color
   Text get violet50 => color(FxColor.violet50);
+
+  /// Returns Text with violet100 color
   Text get violet100 => color(FxColor.violet100);
+
+  /// Returns Text with violet200 color
   Text get violet200 => color(FxColor.violet200);
+
+  /// Returns Text with violet300 color
   Text get violet300 => color(FxColor.violet300);
+
+  /// Returns Text with violet400 color
   Text get violet400 => color(FxColor.violet400);
+
+  /// Returns Text with violet500 color
   Text get violet500 => color(FxColor.violet500);
+
+  /// Returns Text with violet600 color
   Text get violet600 => color(FxColor.violet600);
+
+  /// Returns Text with violet700 color
   Text get violet700 => color(FxColor.violet700);
+
+  /// Returns Text with violet800 color
   Text get violet800 => color(FxColor.violet800);
+
+  /// Returns Text with violet900 color
   Text get violet900 => color(FxColor.violet900);
 }