+++ lib/src/flutter/mixins/colors.dart
@@ -1,109 +1,265 @@
 import 'package:flutter/material.dart';
 
-/// Custom color palette
+/// Custom color palette for easy access to predefined color options
 mixin FxColor {
+  /// Returns primary color `0xFF006EFF`
   static const Color primary = Color(0xFF006EFF);
+
+  /// Returns secondary color `0xFFE4E6EF`
   static const Color secondary = Color(0xFFE4E6EF);
+
+  /// Returns success color `0xFF2ED477`
   static const Color success = Color(0xFF2ED477);
+
+  /// Returns info color `0xFF8950FC`
   static const Color info = Color(0xFF8950FC);
+
+  /// Returns warning color `0xFFFF9138`
   static const Color warning = Color(0xFFFF9138);
+
+  /// Returns danger color `0xFFF46363`
   static const Color danger = Color(0xFFF46363);
 
+  /// Returns light color `0xFFF3F6F9`
   static const Color light = Color(0xFFF3F6F9);
+
+  /// Returns dark color `0xFF222428`
   static const Color dark = Color(0xFF222428);
 
+  /// Returns white color `0xFFFFFFFF`
   static const Color white = Color(0xFFFFFFFF);
+
+  /// Returns black color `0xFF000000`
   static const Color black = Color(0xFF000000);
 
+  /// Returns Text color `0xFF565360`
   static const Color kcText = Color(0xFF565360);
+
+  /// Returns Label color `0xFF78858`
   static const Color kcLabel = Color(0xFF78858F);
+
+  /// Returns Disabled color `0xFFE1DFE9`
   static const Color kcDisabled = Color(0xFFE1DFE9);
+
+  /// Returns TextField color `0xFFF3F3F`
   static const Color kcTextField = Color(0xFFF3F3F3);
 
-  // Status Colors
-  static const Color kcWarningStatus = Color(0xFFF9AE3F);
-  static const Color kcDangerStatus = Color(0xFFDE5050);
-  static const Color kcSuccessStatus = Color(0xFF37BC64);
+  /// Returns Warning Light color `0xFFFFF8DE`
+  static const Color kcWarningLight = Color(0xFFFFF8DE);
 
-  // Light Status Colors
-  static const Color kcWarningLightStatus = Color(0xFFFFF8DE);
-  static const Color kcDangerLightStatus = Color(0xFFFFF3F3);
-  static const Color kcSuccessLightStatus = Color(0xFFE2F8E9);
+  /// Returns Danger Light color `0xFFFFF3F3`
+  static const Color kcDangerLight = Color(0xFFFFF3F3);
 
+  /// Returns Success Light color `0xFFE2F8E9`
+  static const Color kcSuccessLight = Color(0xFFE2F8E9);
+
+  /// Returns gray50 color `0XFFf9fafb`
   static const Color gray50 = Color(0XFFf9fafb);
+
+  /// Returns gray100 color `0XFFf3f4f6`
   static const Color gray100 = Color(0XFFf3f4f6);
+
+  /// Returns gray200 color `0XFFe5e7eb`
   static const Color gray200 = Color(0XFFe5e7eb);
+
+  /// Returns gray300 color `0XFFd1d5db`
   static const Color gray300 = Color(0XFFd1d5db);
+
+  /// Returns gray400 color `0XFF9ca3af`
   static const Color gray400 = Color(0XFF9ca3af);
+
+  /// Returns gray500 color `0XFF6b7280`
   static const Color gray500 = Color(0XFF6b7280);
+
+  /// Returns gray600 color `0XFF4b5563`
   static const Color gray600 = Color(0XFF4b5563);
+
+  /// Returns gray700 color `0XFF374151`
   static const Color gray700 = Color(0XFF374151);
+
+  /// Returns gray800 color `0XFF1f2937`
   static const Color gray800 = Color(0XFF1f2937);
+
+  /// Returns gray900 color `0XFF111827`
   static const Color gray900 = Color(0XFF111827);
 
+  /// Returns Color color `0XFFfafafa`
   static const Color neutral50 = Color(0XFFfafafa);
+
+  /// Returns Color color `0XFFf5f5f5`
   static const Color neutral100 = Color(0XFFf5f5f5);
+
+  /// Returns Color color `0XFFe5e5e5`
   static const Color neutral200 = Color(0XFFe5e5e5);
+
+  /// Returns Color color `0XFFd4d4d4`
   static const Color neutral300 = Color(0XFFd4d4d4);
+
+  /// Returns Color color `0XFFa3a3a3`
   static const Color neutral400 = Color(0XFFa3a3a3);
+
+  /// Returns Color color `0XFF737373`
   static const Color neutral500 = Color(0XFF737373);
+
+  /// Returns Color color `0XFF525252`
   static const Color neutral600 = Color(0XFF525252);
+
+  /// Returns Color color `0XFF404040`
   static const Color neutral700 = Color(0XFF404040);
+
+  /// Returns Color color `0XFF262626`
   static const Color neutral800 = Color(0XFF262626);
+
+  /// Returns Color color `0XFF171717`
   static const Color neutral900 = Color(0XFF171717);
 
+  /// Returns amber50 color `0XFFfffbeb`
   static const Color amber50 = Color(0XFFfffbeb);
+
+  /// Returns Color color `0XFFfef3c7`
   static const Color amber100 = Color(0XFFfef3c7);
+
+  /// Returns Color color `0XFFfde68a`
   static const Color amber200 = Color(0XFFfde68a);
+
+  /// Returns Color color `0XFFfcd34d`
   static const Color amber300 = Color(0XFFfcd34d);
+
+  /// Returns Color color `0XFFfbbf24`
   static const Color amber400 = Color(0XFFfbbf24);
+
+  /// Returns Color color `0XFFf59e0b`
   static const Color amber500 = Color(0XFFf59e0b);
+
+  /// Returns Color color `0XFFd97706`
   static const Color amber600 = Color(0XFFd97706);
+
+  /// Returns Color color `0XFFb45309`
   static const Color amber700 = Color(0XFFb45309);
+
+  /// Returns Color color `0XFF92400e`
   static const Color amber800 = Color(0XFF92400e);
+
+  /// Returns Color color `0XFF78350f`
   static const Color amber900 = Color(0XFF78350f);
 
+  /// Returns red50 color `0XFFfef2f2`
   static const Color red50 = Color(0XFFfef2f2);
+
+  /// Returns red100 color `0XFFfee2e2`
   static const Color red100 = Color(0XFFfee2e2);
+
+  /// Returns red200 color `0XFFfecaca`
   static const Color red200 = Color(0XFFfecaca);
+
+  /// Returns red300 color `0XFFfca5a5`
   static const Color red300 = Color(0XFFfca5a5);
+
+  /// Returns red400 color `0XFFf87171`
   static const Color red400 = Color(0XFFf87171);
+
+  /// Returns red500 color `0XFFef4444`
   static const Color red500 = Color(0XFFef4444);
+
+  /// Returns red600 color `0XFFdc2626`
   static const Color red600 = Color(0XFFdc2626);
+
+  /// Returns red700 color `0XFFb91c1c`
   static const Color red700 = Color(0XFFb91c1c);
+
+  /// Returns red800 color `0XFF991b1b`
   static const Color red800 = Color(0XFF991b1b);
+
+  /// Returns red900 color `0XFF7f1d1d`
   static const Color red900 = Color(0XFF7f1d1d);
 
+  /// Returns green50 color `0XFFf0fdf4`
   static const Color green50 = Color(0XFFf0fdf4);
+
+  /// Returns Color color `0XFFdcfce7`
   static const Color green100 = Color(0XFFdcfce7);
+
+  /// Returns Color color `0XFFbbf7d0`
   static const Color green200 = Color(0XFFbbf7d0);
+
+  /// Returns Color color `0XFF86efac`
   static const Color green300 = Color(0XFF86efac);
+
+  /// Returns Color color `0XFF4ade80`
   static const Color green400 = Color(0XFF4ade80);
+
+  /// Returns Color color `0XFF22c55e`
   static const Color green500 = Color(0XFF22c55e);
+
+  /// Returns Color color `0XFF16a34a`
   static const Color green600 = Color(0XFF16a34a);
+
+  /// Returns Color color `0XFF15803d`
   static const Color green700 = Color(0XFF15803d);
+
+  /// Returns Color color `0XFF166534`
   static const Color green800 = Color(0XFF166534);
+
+  /// Returns Color color `0XFF14532d`
   static const Color green900 = Color(0XFF14532d);
 
+  /// Returns blue50 color `0XFFeff6ff`
   static const Color blue50 = Color(0XFFeff6ff);
+
+  /// Returns blue100 color `0XFFdbeafe`
   static const Color blue100 = Color(0XFFdbeafe);
+
+  /// Returns blue200 color `0XFFbfdbfe`
   static const Color blue200 = Color(0XFFbfdbfe);
+
+  /// Returns blue300 color `0XFF93c5fd`
   static const Color blue300 = Color(0XFF93c5fd);
+
+  /// Returns blue400 color `0XFF60a5fa`
   static const Color blue400 = Color(0XFF60a5fa);
+
+  /// Returns blue500 color `0XFF3b82f6`
   static const Color blue500 = Color(0XFF3b82f6);
+
+  /// Returns blue600 color `0XFF2563eb`
   static const Color blue600 = Color(0XFF2563eb);
+
+  /// Returns blue700 color `0XFF1d4ed8`
   static const Color blue700 = Color(0XFF1d4ed8);
+
+  /// Returns blue800 color `0XFF1e40af`
   static const Color blue800 = Color(0XFF1e40af);
+
+  /// Returns blue900 color `0XFF1e3a8a`
   static const Color blue900 = Color(0XFF1e3a8a);
 
+  /// Returns Color color `0XFFf5f3ff`
   static const Color violet50 = Color(0XFFf5f3ff);
+
+  /// Returns Color color `0XFFede9fe`
   static const Color violet100 = Color(0XFFede9fe);
+
+  /// Returns Color color `0XFFddd6fe`
   static const Color violet200 = Color(0XFFddd6fe);
+
+  /// Returns Color color `0XFFc4b5fd`
   static const Color violet300 = Color(0XFFc4b5fd);
+
+  /// Returns Color color `0XFFa78bfa`
   static const Color violet400 = Color(0XFFa78bfa);
+
+  /// Returns Color color `0XFF8b5cf6`
   static const Color violet500 = Color(0XFF8b5cf6);
+
+  /// Returns Color color `0XFF7c3aed`
   static const Color violet600 = Color(0XFF7c3aed);
+
+  /// Returns Color color `0XFF6d28d9`
   static const Color violet700 = Color(0XFF6d28d9);
+
+  /// Returns Color color `0XFF5b21b6`
   static const Color violet800 = Color(0XFF5b21b6);
+
+  /// Returns Color color `0XFF4c1d95`
   static const Color violet900 = Color(0XFF4c1d95);
 }